[DEBUG] Fix build on GCC 7 PC platform
Add -Wno-error in CFLAGS.
This fixes the build on GCC 7.1.1 where the implicit-fallthrough
warning appears to be enabled by default.
The unwanted side effect is that the compiler will not error out on
more serious warnings.

This commit is a copy of the one made by
Hugo Grostabussiat on the json standalone repo.

Change-Id: I7a2198c77a23532d1ed0fdf2d77d4bce4bfff73b
Reviewed-on: https://gerrit-doc.parrot.biz/80764
Reviewed-by: Florent CHARRAIRE <florent.charraire@parrot.com>
Reviewed-by: Yves-Marie Morgan <yves-marie.morgan@parrot.com>